+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [api] Allow access to test coverage reports when not in production
 - [api] Added middleware to guard routes in ephemeral mode
 - [api] Add curl to dev container image
+- [api] Guard routes while reconfiguring
 - [core] Guard routes while reconfiguring
 - [core] Setup unit testing for core
 - [core] Added middleware to guard routes in ephemeral mode